+++ raffia/src/parser/builder.rs
@@ -20,6 +20,7 @@ pub struct ParserBuilder<'cmt, 's: 'cmt> {
 impl<'cmt, 's: 'cmt> ParserBuilder<'cmt, 's> {
     /// Create a parser builder from given source code.
     pub fn new(source: &'s str) -> Self {
+        let source = source.strip_prefix('\u{feff}').unwrap_or(source);
         ParserBuilder {
             source,
             options: None,
@@ -57,9 +58,8 @@ impl<'cmt, 's: 'cmt> ParserBuilder<'cmt, 's> {
 
     /// Build a parser.
     pub fn build(self) -> Parser<'cmt, 's> {
-        let source = self.source.strip_prefix('\u{feff}').unwrap_or(self.source);
         Parser {
-            source,
+            source: self.source,
             syntax: self.syntax,
             options: self.options.unwrap_or_default(),
             tokenizer: Tokenizer::new(self.source, self.syntax, self.comments),