+++ CHANGELOG.md
@@ -3,6 +3,7 @@ _This changelog only relates to the `raffia` crate._
 ## Unreleased
 
 - Allowed comments in compound selector.
+- Fixed stripping UTF-8 BOM when creating parser with `ParserBuilder`.
 
 ## v0.6.0