+++ CHANGELOG.md
@@ -6,6 +6,7 @@ _This changelog only relates to the `raffia` crate._
 - Fixed stripping UTF-8 BOM when creating parser with `ParserBuilder`.
 - Fixed possible infinite loop when parsing pseudo selector arguments.
 - Fixed line break in Sass interpolated URL.
+- Fixed line comments in Sass (indented syntax).
 
 ## v0.6.0