+++ CHANGELOG.md
@@ -2,6 +2,8 @@ _This changelog only relates to the `raffia` crate._
 
 ## Unreleased
 
+## v0.7.0
+
 - Allowed comments in compound selector.
 - Fixed stripping UTF-8 BOM when creating parser with `ParserBuilder`.
 - Fixed possible infinite loop when parsing pseudo selector arguments.